+++ collective/elasticsearch/browser/search.py
@@ -1,3 +1,4 @@
+# -*- coding: utf-8 -*-
 from Products.CMFPlone.browser import search