+++ collective/elasticsearch/indexes.py
@@ -1,5 +1,4 @@
-from logging import getLogger
-
+# -*- coding: utf-8 -*-
 from Missing import MV
 from Acquisition import aq_parent, aq_base
 from DateTime import DateTime