+++ collective/elasticsearch/indexes.py
@@ -29,6 +29,8 @@ def _one(val):
 def _zdt(val):
     if type(val) == datetime:
         val = DateTime(val)
+    elif isinstance(val, basestring):
+        val = DateTime(val)
     return val
 
 
@@ -139,7 +141,7 @@ class EDateIndex(BaseIndex):
             return {'range': {name: {'gte': first}}}
         elif range_ == 'max':
             return {'range': {name: {'lte': first}}}
-        elif range_ == 'min:max' and type(query) in (list, tuple) and \
+        elif range_ == 'minmax' and type(query) in (list, tuple) and \
                 len(query) == 2:
             return {
                 'and': [