+++ collective/elasticsearch/indexes.py
@@ -196,7 +196,7 @@ class EZCTextIndex(BaseIndex):
 
     def get_query(self, name, value):
         value = self._normalize_query(value)
-        clean_value = value.strip('*')  # el doesn't care about * like zope catalog does
+        clean_value = value.strip('*') if value else ""  # el doesn't care about * like zope catalog does
         queries = [{"match_phrase": {name: {
             'query': clean_value,
             'slop': 2